requirements.txt
ib_insync
pandas
numpy
python-dotenv
vectorbt
pytest
flake8
sqlite-utils
discord-webhook
pyyaml
python-dotenv
pytest
pandas
